+++ src/whatsapp/repository/auth.repository.ts
@@ -4,7 +4,7 @@ import { IInsert, Repository } from '../abstract/abstract.repository';
 import { IAuthModel, AuthRaw } from '../models';
 import { readFileSync } from 'fs';
 import { AUTH_DIR } from '../../config/path.config';
-import { InstanceDto } from '../dto/instance.dto';
+import { Logger } from '../../config/logger.config';
 
 export class AuthRepository extends Repository {
   constructor(
@@ -16,24 +16,35 @@ export class AuthRepository extends Repository {
   }
 
   private readonly auth: Auth;
+  private readonly logger = new Logger('AuthRepository');
 
   public async create(data: AuthRaw, instance: string): Promise<IInsert> {
     try {
+      this.logger.verbose('creating auth');
       if (this.dbSettings.ENABLED) {
+        this.logger.verbose('saving auth to db');
         const insert = await this.authModel.replaceOne(
           { _id: instance },
           { ...data },
           { upsert: true },
         );
+
+        this.logger.verbose('auth saved to db: ' + insert.modifiedCount + ' auth');
         return { insertCount: insert.modifiedCount };
       }
+
+      this.logger.verbose('saving auth to store');
 
       this.writeStore<AuthRaw>({
         path: join(AUTH_DIR, this.auth.TYPE),
         fileName: instance,
         data,
       });
+      this.logger.verbose(
+        'auth saved to store in path: ' + join(AUTH_DIR, this.auth.TYPE) + '/' + instance,
+      );
 
+      this.logger.verbose('auth created');
       return { insertCount: 1 };
     } catch (error) {
       return { error } as any;
@@ -42,9 +53,13 @@ export class AuthRepository extends Repository {
 
   public async find(instance: string): Promise<AuthRaw> {
     try {
+      this.logger.verbose('finding auth');
       if (this.dbSettings.ENABLED) {
+        this.logger.verbose('finding auth in db');
         return await this.authModel.findOne({ _id: instance });
       }
+
+      this.logger.verbose('finding auth in store');
 
       return JSON.parse(
         readFileSync(join(AUTH_DIR, this.auth.TYPE, instance + '.json'), {